tustus_2.6.1/requirements.txt
python-telegram-bot[job-queue]==21.6
requests
beautifulsoup4
lxml
tenacity
python-dotenv
httpx~=0.27
